Refactor and improve integration tests suite
